feat(allPokemon): Rearranges files and calls all pokemon to the front end
